+++ ft_memcmp.c
@@ -6,17 +6,18 @@
 /*   By: ayzahrao <ayzahrao@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 17:21:28 by ayzahrao          #+#    #+#             */
-/*   Updated: 2024/01/03 16:03:34 by ayzahrao         ###   ########.fr       */
+/*   Updated: 2024/01/18 06:17:01 by ayzahrao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_memcmp(const void *s1, const void *s2, unsigned int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned char *x;
-	unsigned char *y;
-	unsigned int i;
+	unsigned char	*x;
+	unsigned char	*y;
+	unsigned int	i;
+
 	x = (unsigned char *)s1;
 	y = (unsigned char *)s2;
 	i = 0;
